requirements.txt
discord.py==2.0.0
Flask==2.2.2
